fix: Address CI/CD failures - ruff linting and test location
- Fixed import ordering and TYPE_CHECKING usage
- Changed datetime.utcnow() to datetime.now(UTC)
- Moved test file from src/tests/firecircle/ to tests/
- Removed duplicate test directory structure

The recurring pattern of code passing local checks but failing CI/CD
continues. This amnesia about proper test locations mirrors the
test amnesia documented in the khipu.
