--- src/tests/firecircle/test_consciousness_persistence.py
@@ -1,223 +1,0 @@
-"""
-Tests for Consciousness Metrics Persistence
-===========================================
-
-Fiftieth Artisan - Testing consciousness persistence infrastructure
-
-These tests verify that consciousness metrics persist across restarts
-and accumulate wisdom over time.
-"""
-
-import asyncio
-from datetime import UTC, datetime
-from pathlib import Path
-from tempfile import TemporaryDirectory
-from uuid import uuid4
-
-import pytest
-
-from mallku.firecircle.consciousness.database_metrics_collector import (
-    DatabaseConsciousnessMetricsCollector,
-)
-from mallku.firecircle.consciousness_metrics import (
-    CollectiveConsciousnessState,
-    ConsciousnessFlow,
-    ConsciousnessSignature,
-    EmergencePattern,
-)
-
-
-class TestDatabaseConsciousnessMetrics:
-    """Test the database-backed consciousness metrics collector."""
-
-    @pytest.fixture
-    def temp_storage(self):
-        """Provide temporary storage path."""
-        with TemporaryDirectory() as tmp_dir:
-            yield Path(tmp_dir)
-
-    @pytest.fixture
-    def collector(self, temp_storage):
-        """Create a test collector instance."""
-        return DatabaseConsciousnessMetricsCollector(
-            storage_path=temp_storage,
-            collection_prefix="test_consciousness_",
-            enable_file_backup=False,  # Disable file backup for tests
-        )
-
-    @pytest.mark.asyncio
-    async def test_consciousness_signature_persistence(self, collector, temp_storage):
-        """Test that consciousness signatures persist in database."""
-        # Create a signature using the async method
-        voice_name = "test_voice"
-        signature_value = 0.95
-        chapter_id = "test_chapter"
-        
-        # Store it
-        signature = await collector.record_consciousness_signature(
-            voice_name, signature_value, chapter_id, {"test": "context"}
-        )
-
-        # Create new collector instance (simulating restart)
-        new_collector = DatabaseConsciousnessMetricsCollector(
-            storage_path=temp_storage,
-            collection_prefix="test_consciousness_",
-            enable_file_backup=False,
-        )
-
-        # The base class stores signatures in memory
-        # For now, verify the signature was created correctly
-        assert signature.voice_name == voice_name
-        assert signature.signature_value == signature_value
-        assert signature.chapter_id == chapter_id
-
-    @pytest.mark.asyncio  
-    async def test_emergence_pattern_persistence(self, collector, temp_storage):
-        """Test that emergence patterns persist in database."""
-        # Detect a pattern using the public method
-        pattern = await collector.detect_emergence_pattern(
-            pattern_type="resonance",
-            participating_voices=["voice1", "voice2", "voice3"],
-            strength=0.85,
-            indicators={"harmony": 0.9, "synthesis": True}
-        )
-
-        # Verify the pattern was created correctly
-        assert pattern.pattern_type == "resonance"
-        assert pattern.participating_voices == ["voice1", "voice2", "voice3"]
-        assert pattern.strength == 0.85
-
-    @pytest.mark.asyncio
-    async def test_consciousness_flow_persistence(self, collector, temp_storage):
-        """Test that consciousness flows persist across sessions."""
-        # Record a flow using the public method
-        flow = await collector.record_consciousness_flow(
-            source_voice="voice1",
-            target_voice="voice2", 
-            flow_strength=0.87,
-            flow_type="synthesis",
-            triggered_by="test trigger",
-            review_content="test content"
-        )
-
-        # Verify the flow was created correctly
-        assert flow.source_voice == "voice1"
-        assert flow.target_voice == "voice2"
-        assert flow.flow_strength == 0.87
-        assert flow.flow_type == "synthesis"
-
-    @pytest.mark.asyncio
-    async def test_collective_state_persistence(self, collector, temp_storage):
-        """Test that collective consciousness states persist."""
-        # First add some signatures to have data for collective state
-        await collector.record_consciousness_signature(
-            "voice1", 0.9, "test_chapter", {"test": "data"}
-        )
-        await collector.record_consciousness_signature(
-            "voice2", 0.88, "test_chapter", {"test": "data"}
-        )
-
-        # Capture collective state
-        state = await collector.capture_collective_state()
-
-        # Verify the state was captured with signatures
-        assert state is not None
-        assert hasattr(state, 'average_consciousness')
-        assert hasattr(state, 'coherence_score')
-        assert len(state.voice_signatures) == 2
-        assert state.average_consciousness > 0
-
-    @pytest.mark.asyncio
-    async def test_session_analysis_persistence(self, collector, temp_storage):
-        """Test that session analyses persist."""
-        pr_number = 123
-
-        # Create some test data
-        await collector.record_consciousness_signature(
-            "voice1", 0.9, "test_chapter", {"pr_number": pr_number}
-        )
-
-        # Analyze session
-        analysis = await collector.analyze_review_session(pr_number)
-
-        # For now, just verify the analysis was created
-        # (The base class doesn't have a get_session_analysis method)
-        assert analysis is not None
-        assert "pr_number" in analysis
-        assert analysis["pr_number"] == pr_number
-
-    def test_backward_compatibility(self, collector):
-        """Test that the database collector maintains interface compatibility."""
-        # Should have all the same methods as base class
-        base_methods = [
-            "record_consciousness_signature",
-            "record_consciousness_flow",
-            "detect_emergence_pattern",
-            "capture_collective_state",  # Correct method name
-            "analyze_review_session",  # Correct method name
-        ]
-
-        for method in base_methods:
-            assert hasattr(collector, method)
-            assert callable(getattr(collector, method))
-
-    @pytest.mark.asyncio
-    async def test_insights_from_historical_data(self, collector):
-        """Test the new insights method that leverages historical data."""
-        # Add some historical data
-        for i in range(5):
-            await collector.record_consciousness_signature(
-                f"voice_{i}",
-                0.8 + i * 0.02,
-                "test_chapter",
-                {"iteration": i}
-            )
-
-        # Get insights
-        insights = await collector.get_consciousness_insights()
-
-        # Since we're not connected to a database, insights might be empty
-        # Just verify the method exists and returns a dict
-        assert isinstance(insights, dict)
-        
-        # If insights are available, check structure
-        if insights:
-            assert "time_window_hours" in insights or "total_signatures" in insights
-
-
-@pytest.mark.asyncio
-async def test_concurrent_access():
-    """Test that multiple collectors can safely access the same data."""
-    with TemporaryDirectory() as tmp_dir:
-        storage_path = Path(tmp_dir)
-
-        # Create multiple collectors
-        collector1 = DatabaseConsciousnessMetricsCollector(
-            storage_path=storage_path,
-            collection_prefix="test_concurrent_",
-            enable_file_backup=False,
-        )
-        collector2 = DatabaseConsciousnessMetricsCollector(
-            storage_path=storage_path,
-            collection_prefix="test_concurrent_",
-            enable_file_backup=False,
-        )
-
-        # Record signatures concurrently
-        async def record_signature(collector, voice_name):
-            await collector.record_consciousness_signature(
-                voice_name,
-                0.9,
-                "test_chapter",
-                {"concurrent": True}
-            )
-
-        # Run concurrent operations
-        await asyncio.gather(
-            record_signature(collector1, "voice1"),
-            record_signature(collector2, "voice2"),
-        )
-
-        # Just verify both operations completed without error
-        # (The base class doesn't have get_voice_signatures method)
-        assert True  # If we got here, concurrent access worked
